+++ scripts/run_with_fresh_memory.sh
@@ -46,3 +46,5 @@ echo ""
 python -m src.main
 
 
+
+